Add authors and blog categories to sitemap
