+++ data/blogTags.ts
@@ -1,0 +1,10 @@
+export const badgeTypes = {
+  marketing: "primary",
+  engineering: "success",
+  connections: "info",
+  notes: "warning",
+  product: "danger",
+  operations: "secondary",
+  company: "info",
+  sync: "danger",
+};
