+++ pages/blog/author/[author]/index.tsx
@@ -1,5 +1,5 @@
 import BlogIndex, { getStaticProps as blogIndexStaticProps } from "../../index";
-import Authors from "../../../../data/authors";
+import { getAuthorPaths } from "../../../../utils/sitemap";
 
 export default function BlogIndexWithCategory({ pageProps }) {
   return <BlogIndex pageProps={pageProps} />;
@@ -10,10 +10,6 @@ export async function getStaticProps(ctx) {
 }
 
 export async function getStaticPaths() {
-  const paths = Authors.map((a) => `/blog/author/${a.slug}`);
-
-  return {
-    paths,
-    fallback: false,
-  };
+  const { paths, fallback } = await getAuthorPaths();
+  return { paths, fallback };
 }
